Recording state service refactor
